updates package and dependencies. Reorganizes dice methods. Gameplay not yet working--need to wait for user input on turn
