+++ gameplay/diceMethods.test.js
