+++ package.json
@@ -19,17 +19,19 @@
   },
   "homepage": "https://github.com/twzkraus/yahtzee#readme",
   "dependencies": {
+    "express": "^4.17.1",
     "jquery": "^3.5.1",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react-dom": "^16.13.1",
+    "schema-utils": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
+    "@babel/core": "^7.12.3",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
     "babel-loader": "^8.1.0",
     "jest": "^26.4.2",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
+    "webpack": "^5.3.2",
+    "webpack-cli": "^4.1.0"
   }
 }
