+++ src/components/App.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { display, play } from '../../diceMethods.js';
+import { display, play } from '../../gameplay/diceMethods.js';
 import Die from './Die.jsx';
 import ScoreForm from './ScoreForm.jsx';
 
@@ -9,6 +9,17 @@ const App = (props) => {
   const [diceVals, setDiceVals] = useState([1, 1, 1, 1, 1]);
   const [possScores, setPossScores] = useState([]);
 
+  const takeTurn = async () => {
+    let rollsMade = 0;
+    let acceptedHand = [];
+    while (rollsMade < 3 && acceptedHand.length < 5) {
+      let thisRoll = play.rollDice(5 - acceptedHand.length);
+      setDiceVals(thisRoll);
+      setPossScores(display.getScoreOptions(thisRoll));
+      rollsMade++;
+    }
+  };
+
 
   return (
     <div className="mainContent">
@@ -16,7 +27,7 @@ const App = (props) => {
         {diceVals.map((die, i) => <Die val={diceVals[i]} position={i}/>)}
       </div>
       <div id="buttonBox">
-        <button id="rollBtn">Start Turn</button>
+        <button onClick={takeTurn}>Start Turn</button>
       </div>
       <div className="scoreBox">
         <ScoreForm scores={possScores} />
